+++ CHANGELOG.md
@@ -75,6 +75,8 @@
   connectivity issues. Added `NotProgrammed` condition reason to differentiate
   transient failures from permanent reference errors.
   [#3463](https://github.com/Kong/kong-operator/pull/3463)
+- Fix counting of route attached to a listener by taking into account hostname intersection between the listener and the route.
+  [#3490](https://github.com/Kong/kong-operator/pull/3490)
 
 ## [v2.1.1]